Renamed legacy basic translation to basicsOld
