+++ src/i18n/en-US/draftModelPlan/generalCharacteristics.ts
